Fix ternary operator recognition in Cg/HLSL highlighting.
It was recognising declarations with semantic annotations (which start
with a colon, like Perl 5 attributes) where the file in fact contained a
use of the ternary operator. I fixed it by adding the semantic matcher
to the group of things in the C highlighter which can't be inside
ternary expressions.