Add pathogen bundle for 'vim-hoogle'
It's a plugin for getting at Hoogle searches with a command :Hoogle. It
doesn't seem to be very good so far, but we'll see.

I've also updated vimrc to use the pathogen :Helptags to generate tags
for all installed bundles: it seems to be the only sensible way to keep
them up-to-date when submodules are involved.